feat: 添加多个模型字段和服务方法
为多个模型添加新字段，包括订单发票、优惠券使用记录、客服分类等
扩展多个服务接口，添加新方法用于标签管理、地区查询、客服对话统计等
修复产品详情保存方法的递归调用问题
重构客服对话服务方法命名以保持一致性

+++ daily-discover-product/src/main/java/com/dailydiscover/service/impl/UserReviewStatsServiceImpl.java
@@ -24,7 +24,7 @@ public class UserReviewStatsServiceImpl extends ServiceImpl<UserReviewStatsMappe
     
     @Override
     public boolean updateUserReviewStats(Long userId, Integer totalReviews, Integer helpfulReviews, 
-                                        Integer unhelpfulReviews, Double averageRating) {
+                                        Integer unhelpfulReviews, BigDecimal averageRating) {
         UserReviewStats stats = getByUserId(userId);
         if (stats == null) {
             stats = new UserReviewStats();
